Break HA Guide into chapters and sections
This patch breaks the monolithic bk-ha-guide.xml file into chapters
and sections. Section files are placed in subdirectories, with the
subdirectories named after the chapters (and parts) to which they
belong.

This patch just does structural fixes. Once it's in, we can begin to
do content cleanup in manageable chunks.

Change-Id: I27397834141a3e6c305f60e71350ce869ab7c8a1
Implements: blueprint convert-ha-guide-to-docbook